module8-task1: + auth-form component
