+++ src/components/auth-form/common.js
