+++ src/components/auth-screen/auth-screen.jsx
@@ -4,17 +4,11 @@ import {Link} from 'react-router-dom';
 import PropTypes from 'prop-types';
 import {AppRoute} from '~/common/enums/enums';
 import Header from '~/components/header/header';
-import {LoginFormKey} from './common';
+import AuthForm from '~/components/auth-form/auth-form';
 import {UserActionCreator} from '~/store/actions/user/user';
 
 const AuthScreen = ({formState, onFormStateChange}) => {
   const dispatch = useDispatch();
-
-  const handleFormChange = ({target}) => {
-    const {name, value} = target;
-
-    onFormStateChange(name, value);
-  };
 
   const handleSubmit = (evt) => {
     evt.preventDefault();
@@ -29,43 +23,11 @@ const AuthScreen = ({formState, onFormStateChange}) => {
         <div className="page__login-container container">
           <section className="login">
             <h1 className="login__title">Sign in</h1>
-            <form
-              className="login__form form"
-              onSubmit={handleSubmit}
-              action="#"
-              method="post"
-            >
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input
-                  className="login__input form__input"
-                  onChange={handleFormChange}
-                  value={formState.email || ``}
-                  name={LoginFormKey.EMAIL}
-                  type="email"
-                  placeholder="Email"
-                  required
-                />
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input
-                  className="login__input form__input"
-                  onChange={handleFormChange}
-                  value={formState.password || ``}
-                  name={LoginFormKey.PASSWORD}
-                  type="password"
-                  placeholder="Password"
-                  required
-                />
-              </div>
-              <button
-                className="login__submit form__submit button"
-                type="submit"
-              >
-                Sign in
-              </button>
-            </form>
+            <AuthForm
+              formState={formState}
+              onFormSubmit={handleSubmit}
+              onFormStateChange={onFormStateChange}
+            />
           </section>
           <section className="locations locations--login locations--current">
             <div className="locations__item">
